+++ CHANGELOG.rst
@@ -2,7 +2,13 @@
 Changelog for package pylon_ros2_camera
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.0.2 (2023-11-01)
+-------------------
+* All streamable parameters are set, regardless of the loaded user set
+* New streamable parameter set: frame transmission delay
+
 2.0.1 (2023-10-26)
+-------------------
 * Blaze functionalities are now included in the official galactic and humble branches. Beware that it is now requiring the installation of the pylon Supplementary Package for blaze to be able to compile the driver.
 * Readme file updated with new Basler website links.